andar thi akhu check done and put alert done Outside remaining

+++ client/src/app/dashboard/department/[department]/branch/[branch]/page.tsx
@@ -6,7 +6,8 @@ import { LuPencil } from "react-icons/lu";
 import { FaPlus } from "react-icons/fa";
 import { useRouter } from "next/navigation";
 import { FiEdit, FiTrash2 } from "react-icons/fi";
-
+import Alerts from "@/app/dashboard/Alerts";
+import Loading from "@/app/dashboard/loading";
 const page = () => {
   const [uni_id, setUni_id] = useState("");
   const [class_id, setClass_id] = useState(0);
@@ -19,8 +20,15 @@ const page = () => {
       branch_id: 0,
     },
   ]);
+  const [alertData, setAlertData] = useState({
+    status: 0,
+    function_name: "",
+    isModalOpen: false,
+    onConfirm: (confirm: boolean) => {},
+  });
   const router = useRouter();
   const params = useParams();
+  const [loading, setLoading] = useState(true);
   const [isModalOpen, setIsModalOpen] = useState(false);
   const [inputData, setInputData] = useState({
     class_no: 0,
@@ -56,29 +64,73 @@ const page = () => {
     const data = await response.json();
     if (Array.isArray(data.data)) {
       setClassData(data.data);
+      setLoading(false);
     } else {
       setClassData([]);
     }
   };
 
   const handle_delete = async (class_id: any) => {
-    const response = await fetch(
-      `http://localhost:3000/api/university/${uni_id}/department/${dept_id}/branch/${branch_id}/class/${class_id}`,
-      {
-        method: "DELETE",
-        headers: {
-          "Content-Type": "application/json",
-        },
-      }
-    );
-    const data = await response.json();
-    if (data.status === 201) {
-      console.log("Data successfully deleted:", data);
-      setClassData(classData.filter((data) => data.id !== class_id));
-      router.refresh();
-    } else {
-      console.error("Error deleting data:", data);
-    }
+    setAlertData({
+      status: 1,
+      function_name: "delete",
+      isModalOpen: true,
+      onConfirm: async (confirm: boolean) => {
+        if (confirm) {
+          setLoading(true);
+          try {
+            const response = await fetch(
+              `http://localhost:3000/api/university/${uni_id}/department/${dept_id}/branch/${branch_id}/class/${class_id}`,
+              {
+                method: "DELETE",
+                headers: {
+                  "Content-Type": "application/json",
+                },
+              }
+            );
+            const data = await response.json();
+            if (data.status === 201) {
+              setLoading(false);
+              setAlertData({
+                status: 201,
+                function_name: "delete",
+                isModalOpen: true,
+                onConfirm: (confirm: boolean) => {},
+              });
+              console.log("Data successfully deleted:", data);
+              setClassData(classData.filter((data) => data.id !== class_id));
+              setTimeout(() => {
+                setAlertData({
+                  status: 0,
+                  function_name: "",
+                  isModalOpen: false,
+                  onConfirm: () => {},
+                });
+              }, 3000);
+              router.refresh();
+            } else {
+              throw new Error("Error deleting data");
+            }
+          } catch (error) {
+            console.error("Error deleting data:", error);
+            setAlertData({
+              status: 400,
+              function_name: "delete",
+              isModalOpen: true,
+              onConfirm: (confirm: boolean) => {},
+            });
+            setTimeout(() => {
+              setAlertData({
+                status: 0,
+                function_name: "",
+                isModalOpen: false,
+                onConfirm: () => {},
+              });
+            }, 3000);
+          }
+        }
+      },
+    });
   };
 
   const handle_insert = () => {
@@ -113,12 +165,27 @@ const page = () => {
           class_no: inputData.class_no,
           total_batches: inputData.total_batches,
           students_per_batch: inputData.students_per_batch,
+          branch_id: branch_id,
         }),
       });
       const result = await response.json();
       console.log("Data successfully posted:", result);
 
       if (result.function_name === "update_class") {
+        setAlertData({
+          status: 201,
+          function_name: "update",
+          isModalOpen: true,
+          onConfirm: (confirm: boolean) => {},
+        });
+        setTimeout(() => {
+          setAlertData({
+            status: 0,
+            function_name: "",
+            isModalOpen: false,
+            onConfirm: () => {},
+          });
+        }, 3000);
         setClassData((prevClass) =>
           prevClass.map((cls) =>
             cls.id === class_id
@@ -134,6 +201,20 @@ const page = () => {
       }
 
       if (result.function_name === "create_class") {
+        setAlertData({
+          status: 201,
+          function_name: "create",
+          isModalOpen: true,
+          onConfirm: (confirm: boolean) => {},
+        });
+        setTimeout(() => {
+          setAlertData({
+            status: 0,
+            function_name: "",
+            isModalOpen: false,
+            onConfirm: () => {},
+          });
+        }, 3000);
         setClassData((prevClass) => [...prevClass, result.data[0]]);
       }
 
@@ -150,12 +231,10 @@ const page = () => {
     }
   };
 
-  console.log(classData);
-
   const get_class_data = classData.map((data, index) => {
     return (
       <div
-        className="main_content group shadow-md relative justify-center items-center w-full font-bold rounded-sm"
+        className="main_content group shadow-md relative justify-center items-center w-full font-bold rounded-sm cursor-pointer"
         key={index}
         onClick={() => {
           router.push(
@@ -166,6 +245,7 @@ const page = () => {
         <div className="edit_delete opacity-0 group-hover:opacity-100 group-hover:backdrop-blur-md group-hover:bg-gray-900 group-hover:bg-opacity-10 transition-all duration-1000  flex  border-black justify-center items-center h-full p-2 w-full absolute">
           <button
             onClick={(e) => {
+              e.stopPropagation();
               handle_edit(
                 data.class_no,
                 data.total_batches,
@@ -180,6 +260,7 @@ const page = () => {
           </button>
           <button
             onClick={(e) => {
+              e.stopPropagation();
               handle_delete(data.id);
             }}
             className="flex gap-1 hover:text-red-600 p-2"
@@ -213,11 +294,13 @@ const page = () => {
         </button>
 
         {isModalOpen && (
-          <div className="fixed top-0 left-0 right-0 bottom-0 bg-gray-900 bg-opacity-50 flex justify-center items-center">
+          <div className="fixed z-10 top-0 left-0 right-0 bottom-0 bg-gray-900 bg-opacity-50 flex justify-center items-center">
             <div className="bg-white p-6 rounded-lg shadow-lg w-1/3">
               <div className="flex relative">
                 <div className="flex-1 ">
-                  <h1 className="text-lg font-bold mb-4">Add New class</h1>
+                  <h1 className="text-lg font-bold mb-4">
+                    {class_id == 0 ? "Add " : "Edit "} New class
+                  </h1>
                 </div>
                 <div className="absolute right-0">
                   <button
@@ -287,10 +370,14 @@ const page = () => {
                 <div className="group flex relative mt-4">
                   <button
                     className="bg-blue-600 text-white px-4 py-2 rounded-md disabled:opacity-45 disabled:cursor-not-allowed"
-                    disabled={!inputData.class_no||!inputData.total_batches||!inputData.students_per_batch}
+                    disabled={
+                      !inputData.class_no ||
+                      !inputData.total_batches ||
+                      !inputData.students_per_batch
+                    }
                   >
                     <span>Submit</span>
-                  </button> 
+                  </button>
                   {/** Tooltip displayed only when the button is disabled and hovered */}
                   {!inputData && (
                     <span
@@ -307,9 +394,37 @@ const page = () => {
           </div>
         )}
       </div>
-      <div className="grid grid-cols-3 w-full gap-5">
-        {classData.length > 1 ? get_class_data : null}
-      </div>
+      {alertData.isModalOpen && (
+        <Alerts
+          status={alertData.status}
+          isModalOpen={alertData.isModalOpen}
+          function_name={alertData.function_name}
+          onConfirm={(confirm) => {
+            alertData.onConfirm(confirm);
+            setAlertData({
+              status: 0,
+              function_name: "",
+              isModalOpen: false,
+              onConfirm: () => {},
+            });
+          }}
+        ></Alerts>
+      )}
+      {loading ? (
+        <div className="flex justify-center items-center">
+          <Loading />
+        </div>
+      ) : (
+        <div className="grid grid-cols-4 w-full gap-5">
+          {get_class_data.length > 0 ? (
+            get_class_data
+          ) : (
+            <div className="text-lg font-bold text-slate-950">
+              No Branch Found
+            </div>
+          )}
+        </div>
+      )}
     </div>
   );
 };
